[WIP] Fixes to argument recovery and simple function return type analysis
Argument recovery is now aware of register aliasing (eax, rax, ...)
and recovers argument types and names accordingly. Return type is
recovered from the last user of a return value register specified
by a calling convention, in a basic block with a `llvm::ReturnInst`.
If it's an `llvm::StoreInst` (aka a write) we take the type and
use it as the function return type. If no such write is found,
the function returns `void`. If more than one return type candidate
is found, we terminate. This should be enhanced in the future.